Separate out key logic loading into separate file per type
