+++ src/Tmds.Ssh/PrivateKeyCredential.cs
@@ -2,13 +2,14 @@
 // See file LICENSE for full license details.
 
 using System;
+using System.Text;
 
 namespace Tmds.Ssh;
 
-public sealed class PrivateKeyCredential : Credential
+public sealed partial class PrivateKeyCredential : Credential
 {
     internal string FilePath { get; }
-    internal string? Passphrase { get; }
+    internal byte[] Passphrase { get; }
 
     public PrivateKeyCredential(string path) : this(path, null)
     { }
@@ -16,6 +17,6 @@ public sealed class PrivateKeyCredential : Credential
     public PrivateKeyCredential(string path, string? passphrase)
     {
         FilePath = path ?? throw new ArgumentNullException(nameof(path));
-        Passphrase = passphrase;
+        Passphrase = string.IsNullOrEmpty(passphrase) ? Array.Empty<byte>() : Encoding.UTF8.GetBytes(passphrase);
     }
 }
